Refactor styles and asset paths; update imports
Moved style definitions from src/components/styles.js to src/assets/css/auth/auth.js and extracted color constants to src/assets/css/general/general.js. Updated asset paths in app.json to use .webp images. Refactored component imports in screens and routes to use new style and utility file locations.
