+++ src/assets/css/general/general.js
@@ -1,0 +1,10 @@
+
+export const Colors = {
+    primary: "#ffffff",
+    secondary: "#E5E7EB",
+    tertiary: "#1F2937",
+    darkLight: "#9CA3AF",
+    brand: "#6D28D9",
+    green: "#10B981",
+    red: "#EF4444",
+};
