+++ src/components/ui/KeyboardAvoiding.js
@@ -1,4 +1,3 @@
-import React from "react";
 import { KeyboardAvoidingView, ScrollView, TouchableWithoutFeedback, Keyboard } from "react-native";
 
 const KeyboardAvoiding = ({children}) =>{
